requirements.txt
Flask==2.3.3
discord.py==2.3.2
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
asyncio==3.4.3
threading2==0.1.2
Flask-Session==0.5.0
requests-oauthlib==1.3.1
